+++ SocialApp/SocialApp/Proxies/CalorieRepositoryProxy.cs
@@ -17,7 +17,7 @@
         }
 
         // Synchronous call using .Result or .GetAwaiter().GetResult()
-        public Calorie GetCaloriesByUserId(int userId)
+        public Calorie GetCaloriesByUserId(long userId)
         {
             // Get the calorie data from the API synchronously
             return this.httpClient.GetFromJsonAsync<Calorie>($"user/{userId}").Result!;
